// task-tracker-backend/src/utils/socket-events.enum.ts
export enum SocketEvents {
  CONNECT_MESSENGER = 'CONNECT_MESSENGER',
  MESSENGER_CONNECTED = 'MESSENGER_CONNECTED',
  REMOVE_MESSENGER = 'REMOVE_MESSENGER',
  MESSENGER_REMOVED = 'MESSENGER_REMOVED',
  CONNECT_TRACKER = 'CONNECT_TRACKER',
  TRACKER_CONNECTED = 'TRACKER_CONNECTED',
  REMOVE_TRACKER = 'REMOVE_TRACKER',
  TRACKER_REMOVED = 'TRACKER_REMOVED',
  GET_JIRA_PROJECTS = 'GET_JIRA_PROJECTS',
  CREATE_JIRA_ISSUE = 'CREATE_JIRA_ISSUE',
  JIRA_ISSUE_CREATED = 'JIRA_ISSUE_CREATED',
  GET_TRELLO_BOARDS = 'GET_TRELLO_BOARDS',
  CREATE_TRELLO_ISSUE = 'CREATE_TRELLO_ISSUE',
  TRELLO_ISSUE_CREATED = 'TRELLO_ISSUE_CREATED',
}